[PW-7331] refactoring
