+++ Block/Adminhtml/Support/ConfigurationSettings.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Adyen\Payment\Block\Adminhtml\Support;
+
+class ConfigurationSettings extends \Magento\Backend\Block\Page
+{
+
+}
