--- Block/Adminhtml/Support/Settings.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Adyen\Payment\Block\Adminhtml\Support;
-
-class Settings extends \Magento\Backend\Block\Page
-{
-
-}
