+++ view/adminhtml/templates/support/configurationsettings.phtml
@@ -1,0 +1,1 @@
+<h1>This is Configuration Settings</h1>
