--- view/adminhtml/templates/support/settings.phtml
@@ -1,1 +1,0 @@
-<h1>This is Settings</h1>
